Buttons Control for Cloud Class Instances Now work
Selecting class only shows menu of class instances. Button for class instance, when selected, mark the cloud instance with a red color. Next step, when class is selected, scale down the alpha value for other classes, not selected.

Note, class instance menu requires that only one is selected at a time. Must be updated if need for multiple to be viewed at the same time